feat: Convert all embeddings from float to double in backend and update tests

+++ src/backend/src/Application/AI/VectorStore/VectorDocument.cs
@@ -4,6 +4,6 @@ public class VectorDocument
 {
     public string Id { get; set; } = null!;
     public string Content { get; set; } = null!;
-    public float[] Vector { get; set; } = null!;
+    public double[] Vector { get; set; } = null!;
     public Dictionary<string, string>? Metadata { get; set; }
 }
