+++ src/backend/src/Application/Common/Interfaces/IEmbeddingProvider.cs
@@ -6,5 +6,5 @@ public interface IEmbeddingProvider
 {
     string ProviderName { get; }
     int MaxTextLength { get; }
-    Task<Result<float[]>> GenerateEmbeddingAsync(string text, CancellationToken cancellationToken = default);
+    Task<Result<double[]>> GenerateEmbeddingAsync(string text, CancellationToken cancellationToken = default);
 }
