+++ src/backend/src/Application/Common/Models/VectorStoreQueryResult.cs
@@ -3,8 +3,8 @@ namespace backend.Application.Common.Models;
 public class VectorStoreQueryResult
 {
     public string Id { get; set; } = null!;
-    public List<float> Embedding { get; set; } = new List<float>();
+    public List<double> Embedding { get; set; } = new List<double>();
     public Dictionary<string, string> Metadata { get; set; } = new Dictionary<string, string>();
-    public float Score { get; set; }
+    public double Score { get; set; }
     public string Content { get; set; } = string.Empty;
 }
